refactor(frontend): component to display collections
